adding filing status code into spouseAndDependent and removing filing status page:
